FIX: Prevent collapsing the root node.
git-svn-id: https://eventghost.svn.sourceforge.net/svnroot/eventghost/branches/bitmonster@1370 18fe9db4-152b-49c1-b228-172182a50ab1

+++ eg/Classes/MainFrame/TreeCtrl.py
@@ -610,6 +610,9 @@ class TreeCtrl(wx.TreeCtrl):
         Handles wx.EVT_TREE_ITEM_COLLAPSING
         """
         itemId = event.GetItem()
+        if itemId == self.GetRootItem():
+            event.Veto()
+            return
         self.TraverseDelete(itemId)
         self.Collapse(itemId)
         self.DeleteChildren(itemId)
